+++ dev-info.md
@@ -21,7 +21,7 @@ Pour litération un on n'a pas besoin de pouvoir modifer les information de comp
 - logout
 
 ### administrator
-- `create a company account`
+- create a company account
 - view interniship list
 - view internship details
 - `add an internship offer`
@@ -34,11 +34,11 @@ Pour litération un on n'a pas besoin de pouvoir modifer les information de comp
 
 ---
 
-## Make the site accessible online
+## `Make the site accessible online`
 
 ---
 
-## Pourquoi est ce que l'étudiant a un mot de passe dans sa table
+## `Pourquoi est ce que l'étudiant a un mot de passe dans sa table`
 
 ---
